requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.16
django-rest-framework==0.1.0
djangorestframework==3.14.0
idna==3.4
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
typing_extensions==4.4.0
urllib3==1.26.14
grpcio-tools ~= 1.30
